Cast size_t for printf

+++ src/lexer.c
@@ -14,7 +14,7 @@ dumpToken(struct VuToken* token) {
     size_t length = token->length > max ? max : token->length;
     memcpy(&buffer, token->content, length);
     buffer[length+1] = '\0';
-    printf("Length: %d\n", length);
+    printf("Length: %llu\n", (unsigned long long)length);
     printf("<Token %d '%s' >\n",
         token->position,
         buffer
